Allow gerrit topic from dirty index
The new branch is based on the current HEAD, so we don't need
to prevent git checkout when the index is not clean.

It's a convenience for people who forget to make a topic until
they're about to commit.

Forward git output to console in gerrit topic call

Return result from cli